src/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
Django==4.1.3
django-crontab==0.7.1
django-extensions==3.2.1
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
docutils==0.19
idna==3.4
inflection==0.5.1
pika==1.3.1
protobuf==4.21.12
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
python-dotenv==0.21.0
pytz==2022.6
rabbitmq==0.2.0
redis==4.5.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.15
